Speed up waitForPlaceholders()
For any test that has an image placeholder, we only check every 10
seconds if the image has been uploaded. This always results in a 10
second delay when running in our CI, which is undesireable - change it
to 1 second initially.

The flipside is that if we have a lot of tests users have left waiting
for placeholders it'll use up server time / db time, so we backoff
exponentially until we're only checking every 30 seconds.