Build jar for releases
Currently we don't actually build the jar for tags. I think if we do
build the jar it'll be available to download from the release pages
which would be handy for people who seem unable to get a working jdk
setup.

Also stop microauth/whitespace jobs running for tags, as there seems
to be no point.