FAPI-RW: Fix response_type=code tests
If the server redirects back, we were looking for the error in the
url fragment, but really it's equally permissible that the server
replies in the url query as that was what was requested.

Allow both by detecting which the server used, mapping the
authorization_endpoint_response in the environment appropriately, and
making sure the conditions use authorization_endpoint_response instead
of callback_params.

closes #487