Allow Conformance Suite to be run/debugged/updated directly in IDE
The current build/debug cycle is a bit laborious, something along the
lines of:

- mvn package
- docker-compose up
- Make changes in IDE
- Switch to command line
- build using mvn
- restart docker container (which takes ~45 seconds on mac on a good day)
- reattach debugger

(or a slight variant using hotswap that sometimes works and but
sometimes fails and requires restarting the jvm)

This commit adds an alternative;

- docker-compose -f docker-compose-dev-mac-nodocker.yml up
- Open IDE and run or debug 'Conformance'
- Make changes in IntelliJ
- Press "Update 'Conformance' Application" in Run menu or Debugger controls
(which should hotswap is possible, and restart the jvm if not)

which takes a few seconds. Running outside of docker is also a lot faster
on mac due to the performance issues docker for mac has when accessing
the .jar from the native filesystem.

A 'Conformance' run configuration for IntelliJ is included and
IntelliJ should automatically pick it up when the project is opened. I
presume something similar is possible for Eclipse don't plan to
investigate that myself.

The essentials of how it works is:

- mongodb is now exposed to host
- apache httpd now passes the requests back to the host

microauth hasn't been considered so you can't use run-tests.sh with
this setup currently.

This is mac specific unfortunately, for two reasons:

i) 'delegated' for mongo in docker compose which is only supported on mac

ii) host.docker.internal in server-nodocker.conf to forward from the
ingress to the java, which is apparently a docker for mac specific thing
(I guess there must be some windows/linux equivalent, you'd think wanting
to access the host from a container would be a common thing, but google
didn't turn up any simple solutions)

The 'nodocker' naming I've used is potentially confusing and better
naming suggestions are welcome. "nodocker" is intended to convey that
the java application is run outside of docker.

We now have an awful lot of docker-compose variants; I suspect we
should retire some in due course.