Rework java implementation of variants
The current implementation tried to define the full behaviour of
variants in the @PublishTestModule annotation, and was limited (it
couldn't make any use of Java language features, hence the version of
accessories based on String names) and was likely to lead to very
large, difficult to follow @PublishTestModule annotations in the future.

This version defines each variant in a separate java method, This
perhaps isn't quite as flexible, but from my experiments refactoring
the CIBA tests to use variants seems to be sufficient for current
needs.

This causes no change to the API/frontend.

There are a few rough edges, and (as per the previous implementation)
you can't use a TestModule that has variants sensibly in a TestPlan
yet.