Remove accessories
Accessories (when used to implement a Variant) are essentially like
native java methods, except they use String for their names.

This mean compile time checking is lost, as is easy refactoring and the
other nice things you get from using language features. If someone
partially renames an accessory, everything compiles and runs fine,
except the application of the accessory is silently ignored.

This makes them very fragile, and in my opinion as currently implemented
they are not suitable for use on a code base that a small-to-medium size
team is working on.

Accessories are necessary for Justin's implementation of variants, so
this change also stops variants working (which will be addressed in a
later commit).