Simplify ConditionSequences
Remove various methods that aren't currently required, aren't documented
and appear to fail silently in the case where a sequence no longer
contains the Condition a test was hoping to replace.

The concept here seems good and it may well be that these are added back
in the future with documentation/unittests/extra robustness, but for now
it seems to be sensible to avoid adding them to the codebase.