Remove 'Default' variant
The user must now always explicitly select a variant.

It was unclear what the use of a Default variant was; I couldn't see a
way to use it in a way that was clear to users what was going on.

If a TestModule does have a sensible default, it seems better if it
explicitly defines the default variant itself - though the use cases
I've come across so far it's always seems better to have a more
explicit name than 'Default'.