Use classes instead of strings to define a test plan's modules
The current method of using a string to define a test plan's
modules prevents compile time checking/refactoring/etc.

Switch to using a list of classes instead. There's no performance
impact as the list of test modules is already cached on first access.

One of the CIBA plans has been converted to the new format as an
example, and the code is structured to allow both methods to exist as
a transition method.

This could probably be further refactored once the old way is removed,
as the code further down is still using the list of strings currently
in places it can likely directly use the class list.