Try defining a stop action for the 'autotest' job
As per:

https://gitlab.com/gitlab-org/gitlab-ce/issues/57820#note_177201207

it seems that if we have multiple deployment jobs (for the same
environment) gitlab only looks at the last one for the on_stop action.

We do need the multiple environments (at least until such point as
there's a way to access the conformance suite API from outside the
cluster; although perhaps there's also a way to make the second one
have a different 'environment' tag, but that may causes other problems).
However it seems that setting 'on_stop' for the 'deployment' job
that runs the tests cures the problem.

Also undo the change made in !418 as it's easier to see the status
this way.