FAPI-CIBA: Implementation of dynamic client registration for ping
This is optional, and in fact we expect most testers to not use it as
dynamic client registration is unlikely to be supported on real world
FAPI-CIBA implementation as they would generally only register trusted
clients.

It is however incredibly useful for testing the conformance suite, as
it means we can create a client with the necessary notification
endpoint registered at the start of the test. (It is essentially
infeasible to pre-setup a client as CIBA only allows a single
notification endpoint per client which would require a huge number of
clients and complication in our CI pipeline.)

closes #389