Ensure FAPI-RW-ID2-OB tests always send & validate an acr
The OpenBanking profile requires that an acr value is always sent (as
does FAPI-RW), and the latest feedback is that this is now actually
required. Current the 'first client' test doesn't send an acr, so
change it to request sca or ca, with sca preferred. (The second client
requests sca only.)

This required fixing ValidateIdTokenACRClaims, which was quite broken
as the author appeared to have misunderstood the specifications. I
renamed it to make it clearing it's comparing against the original
request. A unit test has now been added for it.

closes #505