Change FAPI/OB tests to only send x-fapi-financial-id in one case
This header isn't required since FAPI implementers draft 2, but we
were still sending it in all cases.

- Remove the sending from GenerateResourceEndpointRequestHeaders

- Explicitly send header in one case using AddFAPIFinancialIdToResourceEndpointRequest
(interestingly this was created but isn't currently called anywhere...
and was using a different, duplicate, configuration field than the
currently used code, so clean that up)

- Rename GenerateResourceEndpointRequestHeaders to show it's FAPI specific

- Improve error message in GenerateResourceEndpointRequestHeaders

Issue was noticed whilst looking at the code.