Document OB tests requirement for resourceUrl in frontend
I don't think this was made clear anywhere. Possibly the OB tests
should have a configuration field with a different name, so that this
prompt isn't shown for pure FAPI-RW tests, but that would be a more
intrusive change.

The OB URL structure is mandatory and documented here:

https://openbanking.atlassian.net/wiki/spaces/DZ/pages/1077805207/Read+Write+Data+API+Specification+-+v3.1.2#Read/WriteDataAPISpecification-v3.1.2-ResourceURIPathStructure

Inspired by recent queries from users.