Ensure that conditions always log something
This doesn't quite follow Justin's idea on the issue, as when I
tried that you ended up with the oddity that the AbstractCondition
subclasses only have the 'evaluate' function in them, which is not
part of the Condition interface. However the evaluate function
is where the @Pre/PostCondition annotations are, resulting in
AbstractTestModule looking up annotations something that conforms
to Condition by looking for the 'evaluate' function that's not
part of the interface.

Instead we split the task; the Condition keeps track of whether it
has logged anything, and the TestModule logs a message if the
Condition doesn't.

closes #427