Change CIBA ping tests to run to completion in some failure cases
We want tests to run to completion where possible, particularly as our
CI requires it. Instead of throwing an error (which aborts the test),
call a condition which neatly logs the error and marks the test as failed.

There were also a lot of duplicate implementations of
processNotificationCallback for the poll cases, so resolve that down
into a base implementation that fails the test.