Fix various CIBA ping tests that don't error/cleanup properly
If a ping test that calls the backchannel endpoint expecting a failure
actually gets a success then we need to make sure we don't leave a
ping notification pending that could confuses a later test run when
it gets notified.

Also some of the tests we no longer calling the conditions to check
the backchannel endpoint response, so weren't actually raising a
error to make clear to the user that the response was incorrect.