Add variants for running tests
Changed api (e.g runner, plan) to run the test with a variant which user chose.
Stored the test plan and the test variant into the database instead of the test configurations.
Update UI to add new variant selection box. The displayed variant was got as one field of the test plan's or the test's information.
Update UI to display the variant on the log detail page and the plan detail page.

closes #500