Skip checking state in response
Split ValidateErrorResponseFromAuthorizationEndpoint to two new conditions. The first one used for checking state in the authorization response. The second used for checking the error response.
In some tests (e.g ensure-authorization-request-without-request-object, ensure-request-object-without-state) that did not send state in the request object then it was allowed responses with & without state (but if state is there it should be corrected)

closes #492