Raise fault if testmodules relinquish thread with lock held
TestModules must always release the lock before they finish running on
any particular thread, as if not the next chunk of code that runs may well run
on a different thread and deadlock when it tries to take the lock.

This prevents developers writing TestModules that would fail intermittently,
like the failures seen in #510.

We also make sure that the lock is released in the case where an exception
is thrown from a test module, which makes sure that we don't end up
accumulating blocked threads in the system.

This required giving BackgroundTask access to TestRunnerSupport so it
can obtain and call the TestModule to get the lock.

vaguely related to #688