Revert to sending x-fapi-financial-id in all FAPI-RW tests
This reverts part of the changes in !577, as it turns out the OB
standard requires that the header is still sent in all cases in the
3.1.1 OBUK standard many banks are currently working to.

I considered only applying this change to the FAPI-OB tests, but that
looks considerably more intrusive. Possibly that can be looked at
again when we have a bit more time.

I'm not 100% sure of the reason why the calls to
FAPIGenerateResourceEndpointRequestHeaders are located where they are
in the test, particularly the one very close to the start of the test
in configure(). I suspect it might be worth refactoring them to be
closer to the calls to the resource server at some point.

closes #509