Remove dependency between id token signature & alg checks
FAPIValidateIdTokenSigningAlg has an unnecessary dependency on
ValidateIdTokenSignature running successfully. This can end up
producing confusing errors in the test, as
FAPIValidateIdTokenSigningAlg can cause a test to abort not
because the signing alg is bad but because the id token signature
validation had previously failed.

Just remove the dependency by making FAPIValidateIdTokenSigningAlg
parse the id_token, which is only 2 lines of code.