Add gitlab dependency/license scanning jobs
As per:

https://gitlab.com/help/user/application_security/dependency_scanning/index

This might do something to warn us if any of our dependencies have
security vulnerabilities.

It needs an explicit (and not very documented) tweak to enable java
11, it only seems to be mentioned here and not in the main docs:

https://gitlab.com/gitlab-org/security-products/license-management/blob/master/README.md

The current 'autotest' job is renamed 'test', as both templates put
the job into 'test' by default and I'd rather than ran in the same
stage as the 'autotest' as they're less likely to add additional delays
to pipelines if they're there.

This creates the slightly odd situation that the 'test' job is now
not in the 'test' stage. Possibly we could move it to the test stage
too and do all tests post-deployment.