FAPI-CIBA: Refactor 2nd client code into appropriate module
AbstractFAPICIBA had historically had all the code for dealing with a
second client in it, but really it's only used in FAPICIBA.java, so
refactor it there.

This also removes various conditions from the other tests; it's really
not necessary to do the TLS version/cipher tests or the Accept: header
variant tests multiple times so lift them into FAPICIBA.java too.

Additionally now that AbstractFAPICIBA now doesn't default to go on and
do the second client tests we can remove code from other test modules
that was there just to avoid that happening.