Improve error messages/check in ValidateIdTokenACRClaimAgainstRequest
The error messages were less than helpful for users; e.g.
'Missing or invalid acr value' does not help people figure out
if they returned it in the wrong place or returned an array
instead of a string.

Improve that.

This also changes the test such that any returned acr is verified
to be a string, whether an acr value was returned or not.