Add binding message to all FAPI-CIBA requests
This is essentially what the spec requires:

> shall ensure unique authorization context exists in the authorization
> request or require a binding_message in the authentication request;

as there's no real way the conformance suite can pass 'unique
authorization context' currently.

This may need to be finissed for the OB spec, and it makes the
'with-binding-message' test superfluous (and may mean we need a
'without-binding-message' test at some point).

part of #545