Another improvement to logging when test is aborted
This should make cases like:

https://www.certification.openid.net/log-detail.html?log=abx0GWN7Xt

easier to read. But I still think it's too much; probably what we really
want to do is put the info about why the test stopped in the
'interrupted' log entry, rather than having a failure entry that's separate
from the 'interrupted' entry (in addition to the 'failure' entry for the
actual failing condition).

part of #544