Replace sending of acr=sca, essential=true in OB tests
The way the PSD2 law is written, banks have the ultimate authority to
decide on whether 'strong customer authentication' or merely the weaker
'customer authentication' is applied to any given request.

This means that a request for sca with essential:true may result in
either:

1) an error returned as per OIDCC, if the bank is unwilling to perform
sca.

or:

2) a successful request that returns sca

We currently make a request like this when trying to obtain an access
token for the second client, in order to do certificate binding tests,
so outcome '1' is highly undesireable and results in a complaint
authorization server failing the tests - hence change the test to pass
sca /or/ ca in both cases.

closes #579