Revert part of "Fix bug: validateServerJWKs fails with valid server jwks"
This reverts part of commit 6af27092da1724cecd66b6fd5b4c1432b4c9180e,
which was the fix for #577.

The fix incorrectly affects the test for the client jwks (when testing
an AS) which we do need to have only one private key, so we know which
key to use to sign requests. We revert that part of the change and
update the unit test. (I think there is other code that checks there
is only one key, but a log I looked at on the production server showed
the "Invalid JWKs. Private and public exponent don't match" error and it
wasn't obvious which key it was talking about as the user had
incorrectly supplied multiple keys.)