Make the 3.0.7 version of the tests available for certification
3.1.0 added need tests for refresh tokens and requests with no state
value passed, but unfortunately also introduced the variant change
that meant people could no longer rerun test plans created with 3.0.7
(and normally a test plan created with 3.0.7 would still be okay for
certification for a few weeks, allowing people to rerun tests that
previously failed but they or us have now made fixes for).

Adding new tests and insisting certificators immediately pass them
makes life particularly difficult for them as it is not always
possible to make quick changes to production environments.

The exact period of time these tests will be available for is under
discussion in the OIDF certification team but I've put it as 1 month
since the release after 3.0.7 came out for now.

This adds a new test plan to go alongside the current one, and the new
one only contains the tests that were present in 3.0.7.