Make the 'user rejects authentication' failure clearer
Users seem to frequently not reject the authentication here, resulting
in the test failing. Add more text to the error to try and make this
even clearer.

Also rename it, as it should now only be used in situations where the user
was expected to reject the request (and not in other cases where
access_denied errors are expected).