Make 'CheckForKeyIdInServerJWKs' a 'continue on failure'
It's still an error, but a kid isn't actually required for our tests
to run, so carry on and let the rest of the test run.

(I'm not 100% sure it is really 100% required by the specs to have a
kid if you only have one key, however the specs do essentially require
a kid if you want to do key rotation so I don't see a need to
changes it away from being an error just now.)

Changed as a result of the CIBA test server Petteri@ubisecure has
provided not having kid's.