Allow http 400 (in addition to 401) for invalid_client error
It looks like we interpretted the spec incorrect when first writing
this test - although the spec has a 'MUST' here, this only applies
if the client sent an Authorization: header - which is not the case
for either of the allow FAPI-RW/FAPI-CIBA client authentication
methods. Hence 401 is just a 'MAY' and 400 is permitted as per the
introduction in RFC6749 section 5.2.

closes #606