Clean up background tasks in TestModule.stop(), not TestRunner
If an exception is thrown after fireTestFinished() is called (such
as a test interruption from fireTestSkipped()), the cleanup task
will also be removed - but only if this happens in a background
task.

This makes sure remaining background tasks are cleared if the test
is stopped other than due to an error in a background task, and fixes
the problem of the cleanup task getting dropped when the test is skipped.