Initial attempt at adding FAPI-RW JARM support
I've not yet added any mechanism in the front end to activate this;
the most obvious would presumably be to hook it in as 4 new variants
(4 because of private_key+mtls client auth on each of normal + obuk).

This only deals with the case where openid is supported; FAPI-RW also
allows JARM to be used without scope=openid.

The general principle is that the JARM response is unpacked into
the authorization_endpoint_response, then any places that expect
id_token is returned from the authorization endpoint are disabled.