Use authorization_endpoint_response consistently
Historically, authorization_endpoint_response was added to prevent
test conditions needing to know if they should look for the response
in the url query [callback_query_params] or the url fragment
[callback_params]; there's a few straggling pieces of code that were
never fully updated, which are hopefully all now fixed.

This is necessary for JARM changes that extract to
authorization_endpoint_response and hence needs all the conditions to
be looking in the correct place.