FAPI-RW: Add validation that successful responses only contain expected fields
This check was being added as part of the JARM work to verify the JWT
doesn't contain unexpected things (like id_tokens or sub values), but
it seems like a generally correct check to have.

It's only a warning, as technically the specs allow anything to be
returned and require clients to ignore things they don't expect.

The same as the error response, we explicitly allow session_state as
that's a known extension. We should probably add any other known
extensions, if there are any.