Add missing regeneration of client assertion to code-reuse test
This got lost in the refactoring of MR !752

Since generateNewClientAssertion() is only needed for this test,
I've moved it out of the abstract FAPI base class.

Also a duplicate call to CallProtectedResourceWithBearerTokenAndCustomHeaders
has been removed, which was a hangover from when these checks were
run after testing the second client.