Synchronize TestInstanceEventLog
This is shared with BrowserControl, which runs asynchronously, so
we don't want multiple calls running simultaneously. In particular,
startBlock updates the block ID before logging the 'start block'
message.

We _could_ refactor that method to avoid that particular case, but
I think ensuring log writes are sequential is the better fix.

(The question remains as to whether BrowserControl should be using
the same event log object or not)