FAPI-RW: Reject state being returned in url query
Users can get themselves into a big mess as the test suite can produce
confusing errors if the OP returns the response in the url query and also
fails to return an authorization code for a test that should have succeeded.
(We already check for an authorization code being accidentally returned
in the url query)

Add a new test condition that also looks for 'state' in the url query and
ensures we raise a failure for that.