Resolve review comments on GenerateServerConfiguration
As per:

https://gitlab.com/openid/conformance-suite/merge_requests/779#note_250212540

remove the return value that has the potential to cause future
errors.

Also it seems the discoveryUrl being logged did not actually exist, so
change the log to retrieve the values stored into the environment -
what is in the environment for other conditions to use etc is really
the most important thing to log.