Merge the two sequences for verifying client JWT authentication
ValidateClientAuthenticationWithPrivateKeyJWTAndMTLSHolderOfKey and
OIDCCValidateClientAuthenticationWithPrivateKeyJWT should have
actually been doing the same thing.

The non-OIDCC one was previously checking the client key sizes, which
should have been (and is now) in the test module.

The OIDCC one was the only one that verified the client assertion
signature, which clearly the FAPI one should have been doing - so
combining them fixes the FAPI one to do this.