Handle exceptions the same in testrunner vs testdispatcher
Currently we produce a more helpful message for 'callAndStopOnFailure'
if we catch the exception in TestRunner, saying "The failure <x> means
the test cannot continue. Stopping test." which is helpful to users.

I noticed a test had stopped because of an exception caught in
TestDispatcher, which did not have the helpful message.

Some of the other logic was slightly different; mostly this seems to
be slight difference in ordering that appears unimportant.

The important difference is in deciding to return a redirect in some
cases; that is only appropriate in TestDispatcher where the exception
may have happened during the user visiting a test module endpoint that
returns html. That can't happen via TestRunner so is left in
TestDispatcher (albeit it is refactored to be clearer).