Remove incorrect @UserFacing error from CIBA ping handler
The ping handler is called by the AS under test, so doesn't meet the
definition of 'user facing' - i.e. we shouldn't return HTML to the AS.
(It's not clear it's that useful to return a 500 error with a json
payload either, but it's better than a redirect to the log-detail page.)

Spotted via code inspection when trying to understand the code that
looks for the @UserFacing annotation.