Tidy up error() methods in AbstractCondition
Whilst reading the code to investigate #443, I noticed There were two
error() methods that did not log anything, whereas every other 'error'
method did.

It turned out these methods are only used internally in AbstractCondition,
so make them private, and rename them to reflect that they do not log.

This revealed that one of the places these were called in was failing to
add a log entry, so add the missing log call.

It also turned out all the places are related to failed pre/post conditions,
so further rename to show that and remove the need to pass a boolean saying
so.

Also make sure each block of code uses key/object consistently in it's
messages.