Ensure 'setFinalError()' is called before 'stop()'
If this code was called on a background thread it is possible that the
'stop' call aborts the current thread, meaning 'setFinalError' would
never be called.

I'm not sure this can happen in practice in the current code base (if
it can, I suspect the call from FutureWatcher is the most likely
candidate), but regardless it seems wise to avoid the possibility and
to add comments warning others.