Remove 'fireTestFailure()' from public TestModule interface
This is always called before throwing a TestFailureException, but in
some cases a TestFailureException is thrown without calling
fireTestFailure().

It seems like we should either never call fireTestFailure() and hence
rely on the system catching the TestFailureException and marking the test
as failed, or we should always call fireTestFailure() and hence not
rely on the system.

As we have many instances already where the developer failed to called
fireTestFailure() and is already relying on the system to fail the test,
it feels best to always rely on the system catching the exception and
make developers lives simpler (and simplify the TestModule interface).

As far as I can tell, all the code paths that could throw
TestFailureException will be caught by the system and mark the test
as failed. The paths I know about are:

- test module code called as a result of an API call to TestRunner;
I'm not convinced this can actually happen anymore as the calls I found
all execute via runInBackground()

- test module code called via runInBackground where the exception
will be processed via TestRunner's FutureWatcher

- test module code called as a result of a http call direct to a
test module via TestDispatcher

and all those paths correctly called handleTestInterruptedException which
will call fireTestFailure().

Spotted via code inspection when cleaning up the use of ConditionError
for #443.