WIP: Allow a test module to run with different variants to the plan
The current OIDCC client certifications have a requirement that, when
testing client authentication methods other than client_secret_basic,
the client still have to prove that is also supports
client_secret_basic.

There are two ways we could do this:

1) have the client run the testplan with client_secret_basic, as well as the
testplan for any other client authentication types it wants to test

2) have the client run just the testplan for the client authentication
type it wants to certify with, and have a special test module that the
client knows it has to use client_secret_basic with, and that test
module overrides the client authentication variant to be different
than the plan's.

I'm not convinced '2' is ultimately the best way to write the tests,
but it's what we're currently doing and it seems better to have the
variant overridden completely at the system level instead of the test
module trying to pretend it's running than with a different variant
than it actually is.

reverts commit ea3e66f4c25feba88fffaf691381af3ab1254981 and parts of
the others that added OIDCCClientTestClientSecretBasic.

see #687