Remove duplicate messages when a test fails
Currently when a test stops because of a 'callAndStopOnFailure()' condition
failing, we get 3 log entries:

1) the log from AbstractCondition saying the condition failed

2) the message saying the test can't continue because of that failure

3) the message saying test was interrupted before it could complete

'2' was in fact never intended to be logged (see #443), but is the only
place that we currently tell the user why the test didn't continue. So
we remove message '2', but add a new parameter to stop() so that we can
make message '3' include that explanatory text.

We also now log a nicer error if a TestModule throws a
TestFailureException with just a string in it.