Simplify handleException test for ConditionError
We have two slightly different tests that were trying to test the same
thing. The 'instanceof' test is far easier to read so use that to
cover both cases.

I believe they do different in behaviour if we were to ever throw a
subclass of ConditionError, however we have no such subclasses and
arguably the correct behaviour would be the new behaviour of consistently
treating subclasses of ConditionError as having the same behaviour as
ConditionError itself.