BrowserControl: Avoid nesting TestFailureExceptions
BrowserControl throws a number of TestFailureExceptions, and catchs them
and rethrows wrapped in another TestFailureException. Avoid doing that,
as multiple layers of causes can obscure what is going on.

Also if a non-TestFailureException is caught, preserve the cause when
creating the TestFailureException (inorder to preserve the current
message as well, this required adding some new constructors).