Detect attempts to incorrectly throw a ConditionError
This should catch if a developer accidentally throws a ConditionError,
and add an appropriate message in the log. This avoids the ConditionError
potentially not being logged at all.

This can't detect the case where the developer throws a ConditionError
wrapped in a TestFailureException, but in our codebase prior to this
batch of changes that was a rarer case.

see #443