Fix intermittent unit test failure for requestobject conditions
The tests would fail if the clock ticked over a second boundary
between the code under test running and the assertion running.

Change assertion to a 'isCloseTo'. An equally good option would've
been to use
Instant.now(Clock.fixed(Instant.parse("2018-08-22T10:00:00Z"),
ZoneOffset.UTC)) to fix the time in the unit tests, but the
'isCloseTo' solution seems to be the one used in other unit tests so
go for that for consistency.