Add protection against incorrect testids in TestFailureException
It's not hard to create a TestFailureException accidentally passing
some random string instead of a TestId, which currently results in the
exception silently being ignored.

In at least some cases, we can detect when an incorrect
TestFailureException has been thrown, and where we know the testId so
can add an entry to the test log.

In at least one other case, we don't have any access to the testId and
the best we can do is log an error to the console.