Raise fault if Conditions called without lock held
As mentioned in #688, we don't want conditions to be running unless the
Environment is locked. In this commit we add code that will throw errors
if a test tries to run a condition without having the lock held.

It's not uncommon for a developer to get this wrong, and it's tough to
check at code review time, so adding a runtime check seems sensbile.

There are likely some race conditions in this, because of the issues
that #688 notes about setStatus() not being thread safe.

We also have to skip the new thread checks for
SleepUntilAuthReqExpires; as per the comments we will need to revisit
this as this condition does use the environment so is not safe to call
without the lock held - and equally we don't want to call it with the
lock held as we want to allow conditions to run on other threads
during the sleep. It's likely that making it a function in
AbstractTestModule is the better way to go.

As noted in #688 the lock also isn't held whilst running configure(),
so we need to add an exception for when the test is still in CREATED,
for now. If we fixed this, we could make the environment assert that the
lock is held when it is accessed which might reveal some further
unsafe places.

Note that I am fairly certain this breaks some of the HEART tests,
particularly this part of PlainRS.java:

setStatus(Status.WAITING);

callAndStopOnFailure(CallProtectedResourceWithBearerToken.class);

setStatus(Status.RUNNING);

which will now raise a TestFailureException about
CallProtectedResourceWithBearerToken being run without the lock held.
I have no way to test this code, but my suspicion is that it has the
same issue as SleepUntilAuthReqExpires; i.e. it needs to release the
lock (and move from RUNNING->WAITING) for part (i.e. where it access
the environment) but not all of it's code - in this case I believe it
needs to relinquish the lock only whilst making the actual http
transaction, as other parts of the test module need to run before that
http transaction can complete. I suggest leaving that issue to address
as/when necessary; the HEART tests are not currently actively
used/maintained:

http://lists.openid.net/pipermail/openid-specs-heart/Week-of-Mon-20190930/001710.html