Change Environment so that JsonNull is not treated the same as missing
Lots of test code that checks if a value is missing passes even if the
value is actually explicitly present in the JSON but with a null value.

This is because the Environment was explicitly treating JsonNull the same
(returning null) as if the key wasn't present. This hides information
so isn't great, particularly as most specs at least recommend that
missing values aren't encoded with nulls.

closes #702