BrowserControl: preserve cookies during a running test
The OIDC prompt=login tests require that a server asks the user to
login when prompt=login is passed, even if the user is already logged in.

Currently we don't preserve cookies when using the browser automation,
so the user can never be 'logged in', meaning that we wouldn't actually
test the handling of prompt=none when the automation was in use.

This change makes the built in Selenium used for browser automation
perform the same as an end-user web browser does, in that a second
visit to the authorization endpoint (or any other) within the same
test module will now have the cookies that the OP set during the first
call.

This is necessary so that the any user session at the OP is preserved,
which is important as otherwise the tests for prompt=login and
prompt=none will be unable to test the correct thing.

Unfortunately this equally means that we have the situation where the
OP behaves differently on the second call to the Authorization
Endpoint, which means that the existing JSON automations that expect a
'username' and 'password' field to be present no longer work; a
separate commit will allow the JSON to indicate these fields are
optional.

(This would be a problem if a test module ever needs to trigger multiple
different browser sessions, but that is problematic in the non-automation
case anyway, so doesn't seem to be an issue right now.)