BrowserControl: Allow text entry to be optional
The change to BrowserControl so that it preserves cookies between
automated web interactions within a test module instance means that
the OP is now aware of the user session, and hence doesn't ask the
user to login again.

In the Authlete case (as in the CI) the urls used are identical, so
there's no current mechanism to detect that we're in the "already
logged in" situation and stop the input of the username/password
trying to run.

To cater for this we allow a text entry to be marked as "optional",
meaning that the test will continue  even if the element isn't
present - just adding a log entry so it's relatively clear to users
what has happened if something happens they didn't expect.

There might be other approaches, like making commands depend on page
content or somehow indicating to the browser commands that this the
user would be expected to have an active session, but these all sound
a lot more complex.