Make run-test-plan.py fail if test result isn't one of known values
We've seen at least the 'local' tests returning 'UNKNOWN' for a
result, which the script was ignoring; it seems sensible to be a bit
more paranoid and make sure we don't get an 'UNKNOWN' result.

Failed logs are also not handled exactly how you'd expected, so add a
note about that.